[ALPS03396073] memcfg: fix coverity issue
[Detail]
Cast memtotal to unsigned long long to avoid overflow.

MTK-Commit-Id: 896f6e9ce47a9df0769d604decf18a6b58542ef1

CR-Id: ALPS03396073
Feature: Memory
(cherry picked from commit 9f8e69fb7251f010c6c7b68dc37dbd2963ef3f14)

+++ drivers/misc/mediatek/mem/mtk_memcfg_reserve_info.c
@@ -196,7 +196,7 @@ static unsigned long long get_memtotal(void)
 	unsigned long long memtotal;
 
 	si_meminfo(&i);
-	memtotal = i.totalram << PAGE_SHIFT;
+	memtotal = (unsigned long long)i.totalram << PAGE_SHIFT;
 
 	return memtotal;
 }
